Add signal callback for `post_delete`
In order to ensure schema is dropped when a Tenant is deleted, we need a callback. According to Django docs a `signals` submodule is a good place to put it.

> **Where should this code live?**
>
> Strictly speaking, signal handling and registration code can live anywhere you like, although it’s recommended to avoid the application’s root module and its models module to minimize side-effects of importing code.
>
> In practice, signal handlers are usually defined in a signals submodule of the application they relate to. Signal receivers are connected in the ready() method of your application configuration class. If you’re using the receiver() decorator, simply import the signals submodule inside ready().

Looks like if we pin the callback to the base class it will not be called for Models inheriting from it.