Update test matrix for Postgres & Django/Python compatibility jobs
the original matrix was creating test jobs for all combinations of
Python/Postgres/Django versions but in reality it was always
installing the latest version of Django and only tested
compatibility with different Postgres versions.

With this change the original test matrix is simplified to always
use a single Python version (3.8 atm) + latest Django for Postgres
jobs. Another test job is added with always uses a single
Postgres version (12.2 atm) + a matrix of Python/Django versions
to test compatibility with various Django versions. Here
pip install strictly specifies which version of Django is used!